+++ payjoin-cli/Cargo.toml
@@ -55,5 +55,5 @@ url = { version = "2.5.4", features = ["serde"] }
 
 [dev-dependencies]
 nix = { version = "0.30.1", features = ["aio", "process", "signal"] }
-payjoin-test-utils = { version = "0.0.1" }
+payjoin-test-utils = { version = "0.0.1", features = ["bitcoind"] }
 tempfile = "3.20.0"
