+++ payjoin-ffi/javascript/test-utils/Cargo.toml
@@ -13,6 +13,7 @@ serde_json = "1.0.142"
 tokio = { version = "1.47.1", features = ["full"] }
 
 [dependencies.payjoin-test-utils]
+features = ["bitcoind"]
 path = "../../../payjoin-test-utils"
 
 [build-dependencies]
