+++ payjoin/Cargo.toml
@@ -62,7 +62,7 @@ web-time = "1.1.0"
 
 [dev-dependencies]
 once_cell = "1.21.3"
-payjoin-test-utils = { version = "0.0.1" }
+payjoin-test-utils = { version = "0.0.1", features = ["bitcoind"] }
 tokio = { version = "1.47.1", features = ["full"] }
 tracing = "0.1.41"
